feat: splitt root api into parts, add type for user-api, auth-api
