+++ projects/01-first-project/react-1/src/api/api.ts
@@ -1,58 +1,14 @@
 import { login } from './../redux/auth-reducer';
-import { ProfileType } from './../types/types';
+import { ProfileType, UserType } from './../types/types';
 import axios from 'axios'
 
-const instance = axios.create({
+export const instance = axios.create({
     withCredentials: true,
     baseURL: `https://social-network.samuraijs.com/api/1.0/`,
     headers: {
         "API-KEY": "b14f3ddd-1de3-46f5-9fc3-b0a3d168e1af"
     }
 })
-
-
-export const usersAPI = {
-    getUsers(currentPage = 1, pageSize = 10) {
-        return instance.get(`users?page=${currentPage}&count=${pageSize}`)
-            .then(response => {
-                return response.data
-            })
-    },
-    follow(userId: number) {
-        return instance.post(`follow/${userId}`)
-    },
-    unfollow(userId: number) {
-        return instance.delete(`follow/${userId}`)
-    },
-    getProfile(userId: number) {
-        console.warn('Obsolete method. Please use profileAPI object')
-        return profileAPI.getProfile(userId)
-    }
-}
-
-export const profileAPI = {
-    getProfile(userId: number) {
-        return instance.get(`profile/${userId}`)
-    },
-    getStatus(userId: number) {
-        return instance.get(`profile/status/${userId}`)
-    },
-    updateStatus(status: string) {
-        return instance.put(`profile/status`, {status: status})
-    },
-    savePhoto(photoFile: any) {
-        const formData = new FormData()
-        formData.append("image", photoFile)
-        return instance.put(`profile/photo`, formData, {
-            headers: {
-                'Content-Type': 'multipart/form-data'
-            }
-        })
-    },
-    saveProfile(profile: ProfileType) {
-        return instance.put(`profile`, profile)
-    }
-}
 
 export enum ResultCodesEnum {
     Success = 0,
@@ -63,40 +19,14 @@ export enum ResultCodeForCaptcha {
     CapchaIsRequierd = 10
 }
 
-type MeResponseType = {
-    data: {
-        id: number
-        email: string
-        login: string
-    }
-    resultCode: ResultCodesEnum
+export type GetItemsType = {
+    items: Array<UserType>
+    totalCount: number
+    error: string | null
+}
+
+export type ResponseType<D = {}, RC = ResultCodesEnum> = {
+    data: D
     messages: Array<string>
+    resultCode: RC
 }
-
-type LoginResponseType = {
-    data: {
-        userId: number
-    }
-    resultCode: ResultCodesEnum | ResultCodeForCaptcha
-    messages: Array<string>
-}
-
-export const authAPI = {
-    me() {
-        return instance.get<MeResponseType>(`auth/me`).then(res => res.data)
-    },
-    login(email: string, password: string, rememberMe = false, captcha: null | string = null) {
-        return instance.post<LoginResponseType>(`/auth/login`, {email, password, rememberMe, captcha}).then(res => res.data)
-    },
-    logout() {
-        return instance.delete(`/auth/login`)
-    },
-}
-
-export const securityAPI = {
-    getCaptchaUrl() {
-        return instance.get(`security/get-captcha-url`)
-    },
-}
-
-
