+++ projects/01-first-project/react-1/src/api/auth-api.ts
@@ -1,0 +1,23 @@
+import { instance, ResponseType } from './api.ts'
+
+type MeResponseDataType = {
+    id: number
+    email: string
+    login: string
+}
+
+type LoginResponseDataType = {
+    userId: number
+}
+
+export const authAPI = {
+    me() {
+        return instance.get<ResponseType<MeResponseDataType>>(`auth/me`).then(res => res.data)
+    },
+    login(email: string, password: string, rememberMe = false, captcha: null | string = null) {
+        return instance.post<ResponseType<LoginResponseDataType>>(`/auth/login`, {email, password, rememberMe, captcha}).then(res => res.data)
+    },
+    logout() {
+        return instance.delete(`/auth/login`)
+    },
+}
