+++ projects/01-first-project/react-1/src/api/profile-api.ts
@@ -1,0 +1,26 @@
+import { instance } from './api.ts'
+import { ProfileType } from './../types/types'
+
+export const profileAPI = {
+    getProfile(userId: number) {
+        return instance.get(`profile/${userId}`)
+    },
+    getStatus(userId: number) {
+        return instance.get(`profile/status/${userId}`)
+    },
+    updateStatus(status: string) {
+        return instance.put(`profile/status`, {status: status})
+    },
+    savePhoto(photoFile: any) {
+        const formData = new FormData()
+        formData.append("image", photoFile)
+        return instance.put(`profile/photo`, formData, {
+            headers: {
+                'Content-Type': 'multipart/form-data'
+            }
+        })
+    },
+    saveProfile(profile: ProfileType) {
+        return instance.put(`profile`, profile)
+    }
+}
