+++ projects/01-first-project/react-1/src/api/security-api.ts
@@ -1,0 +1,7 @@
+import {instance} from './api.ts'
+
+export const securityAPI = {
+    getCaptchaUrl() {
+        return instance.get(`security/get-captcha-url`)
+    },
+}
