+++ projects/01-first-project/react-1/src/api/users-api.ts
@@ -1,0 +1,23 @@
+import { AxiosPromise } from 'axios'
+import {GetItemsType, instance, ResponseType } from './api.ts'
+import { profileAPI } from './profile-api.ts'
+
+
+export const usersAPI = {
+    getUsers(currentPage = 1, pageSize = 10) {
+        return instance.get<GetItemsType>(`users?page=${currentPage}&count=${pageSize}`)
+            .then(response => {
+                return response.data
+            })
+    },
+    follow(userId: number) {
+        return instance.post<ResponseType>(`follow/${userId}`).then(res => res.data)
+    },
+    unfollow(userId: number) {
+        return instance.delete(`follow/${userId}`).then(res => res.data) as Promise<ResponseType>
+    },
+    getProfile(userId: number) {
+        console.warn('Obsolete method. Please use profileAPI object')
+        return profileAPI.getProfile(userId)
+    }
+}
